+++ carp_mobile_sensing/lib/domain/serialization.dart
@@ -64,6 +64,7 @@ class FromJsonFactory {
     registerFromJsonFunction("MeasureType", MeasureType.fromJsonFunction);
     registerFromJsonFunction("Measure", Measure.fromJsonFunction);
     registerFromJsonFunction("PeriodicMeasure", PeriodicMeasure.fromJsonFunction);
+    registerFromJsonFunction("AppUsageMeasure", AppUsageMeasure.fromJsonFunction);
 
     //    registerFromJsonFunction("AudioMeasure", AudioMeasure.fromJsonFunction);
     //    registerFromJsonFunction("NoiseMeasure", NoiseMeasure.fromJsonFunction);