+++ carp_webservices/CHANGELOG.md
@@ -1,3 +1,6 @@
+### 0.2.9
+* update to `carp_mobile_sensing` v.0.6.0
+
 ## 0.2.8+3
 * fixed bug in token refresh.